add scratch-mine script to print a users instances
